Move `test_util` crate to `janus_core`
Removes the `test_util` crate, relocating the items it defines to a
`test_util` module in `janus_core`, gated by the `test-util` feature. We
also set `resolver = "2"` in the top-level `Cargo.toml`, because
otherwise crate features get unified between dependencies and
dev-dependencies, which cause the release variants of our binary targets
to be built with the `test-util` feature (and indeed, before this PR, we
were building `tokio` with `test-util` all the time, too!). See [1] for
details.

[1]: https://github.com/rust-lang/cargo/issues/4866
