+++ janus_core/Cargo.toml
@@ -11,7 +11,7 @@ rust-version = "1.58"
 
 [features]
 database = ["bytes", "postgres-protocol", "postgres-types"]
-test-util = []
+test-util = ["assert_matches", "serde_json", "futures", "tracing", "tracing-log", "tracing-subscriber", "tokio/macros", "tokio/sync"]
 
 [dependencies]
 anyhow = "1"
@@ -30,6 +30,10 @@ serde = { version = "1.0.138", features = ["derive"] }
 thiserror = "1.0"
 tokio = { version = "^1.19", features = ["rt"] }
 
-[dev-dependencies]
-assert_matches = "1"
-serde_json = "1.0.82"
+# Dependencies required only if feature "test-util" is enabled
+assert_matches = { version = "1", optional = true }
+serde_json = { version = "1.0.82", optional = true }
+futures = { version = "0.3.21", optional = true }
+tracing = { version = "0.1.34", optional = true }
+tracing-log = { version = "0.1.3", optional = true }
+tracing-subscriber = { version = "0.3", features = ["std", "env-filter", "fmt"], optional = true }
