+++ janus_core/src/test_util/runtime.rs
@@ -1,5 +1,5 @@
+use crate::Runtime;
 use futures::FutureExt;
-use janus_core::Runtime;
 use std::{
     collections::HashMap,
     future::Future,
