+++ janus_server/Cargo.toml
@@ -11,7 +11,11 @@ tokio-console = ["console-subscriber"]
 jaeger = ["tracing-opentelemetry", "opentelemetry-jaeger"]
 otlp = ["tracing-opentelemetry", "opentelemetry-otlp", "opentelemetry-semantic-conventions", "tonic"]
 prometheus = ["opentelemetry-prometheus", "dep:prometheus"]
+<<<<<<< HEAD
 test-util = ["dep:janus_test_util", "dep:lazy_static", "dep:testcontainers"]
+=======
+test-util = ["janus_core/test-util", "lazy_static", "testcontainers"]
+>>>>>>> 8a0ce54 (Move `test_util` crate to `janus_core`)
 
 [dependencies]
 anyhow = "1"
@@ -61,6 +65,10 @@ tracing-subscriber = { version = "0.3", features = ["std", "env-filter", "fmt", 
 url = { version = "2.2.2", features = ["serde"] }
 uuid = { version = "1.1.2", features = ["v4"] }
 warp = { version = "^0.3", features = ["tls"] }
+
+# Optional dependencies enabled by feature "test-util"
+lazy_static = { version = "1", optional = true }
+testcontainers = { version = "0.14.0", optional = true }
 
 [dev-dependencies]
 assert_matches = "1"
