+++ janus_server/src/aggregator/aggregation_job_driver.rs
@@ -803,10 +803,10 @@ mod tests {
         },
         message::{Duration, HpkeConfig, Interval, Nonce, NonceChecksum, Report, Role, TaskId},
         task::VdafInstance,
+        test_util::{
+            install_test_trace_subscriber, run_vdaf, runtime::TestRuntimeManager, MockClock,
+        },
         Runtime,
-    };
-    use janus_test_util::{
-        install_test_trace_subscriber, run_vdaf, runtime::TestRuntimeManager, MockClock,
     };
     use mockito::mock;
     use prio::{
