+++ janus_server/src/binary_utils/job_driver.rs
@@ -200,8 +200,10 @@ where
 mod tests {
     use super::*;
     use crate::{message::AggregationJobId, task::VdafInstance};
-    use janus_core::message::TaskId;
-    use janus_test_util::{install_test_trace_subscriber, runtime::TestRuntimeManager, MockClock};
+    use janus_core::{
+        message::TaskId,
+        test_util::{install_test_trace_subscriber, runtime::TestRuntimeManager, MockClock},
+    };
     use tokio::sync::Mutex;
 
     #[tokio::test]
