+++ monolithic_integration_test/Cargo.toml
@@ -10,7 +10,7 @@ publish = false
 chrono = "0.4.19"
 deadpool-postgres = "0.10.1"
 futures = "0.3.21"
-janus_core = { path = "../janus_core" }
+janus_core = { path = "../janus_core", features = ["test-util"] }
 janus_client = { path = "../janus_client" }
 janus_server = { path = "../janus_server", features = ["test-util"] }
 lazy_static = "1"
@@ -19,7 +19,6 @@ rand = "0.8"
 reqwest = { version = "0.11.11", default-features = false, features = ["rustls-tls"] }
 ring = "0.16.20"
 testcontainers = "0.14.0"
-janus_test_util = { path = "../test_util" }
 tokio = { version = "^1.19", features = ["full", "tracing"] }
 tokio-postgres = { version = "0.7.6", features = ["with-chrono-0_4"] }
 tracing = "0.1.34"
