--- test_util/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "janus_test_util"
-version = "0.1.0"
-edition = "2021"
-license = "MPL-2.0"
-rust-version = "1.58"
-publish = false
-
-[dependencies]
-assert_matches = "1"
-chrono = "0.4"
-futures = "0.3.21"
-janus_core = { path = "../janus_core" }
-prio = "0.8.0"
-rand = "0.8"
-ring = "0.16.20"
-tracing = "0.1.34"
-tracing-log = "0.1.3"
-tracing-subscriber = { version = "0.3", features = ["std", "env-filter", "fmt"] }
-tokio = { version = "^1.19", features = ["rt", "sync"] }
-
-[dev-dependencies]
-tokio = { version = "*", features = ["macros"] }
